requirements.txt

requests
feedparser
beautifulsoup4
newspaper3k

faiss-cpu
sentence-transformers
torch 

google-generativeai

python-dotenv

fastapi
uvicorn[standard]
lxml_html_clean
pydantic-settings